service/boardgame-rules-backend/pyproject.toml
[project]
name = "boardgame-rules-backend"
version = "0.1.0"
description = "Backend for Boardgame Rules Assistant"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "boto3>=1.42.64",
    "celery>=5.6.2",
    "fastapi>=0.135.1",
    "llama-index>=0.14.16",
    "llama-index-postprocessor-sbert-rerank>=0.4.2",
    "llama-index-vector-stores-qdrant>=0.9.2",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2",
    "pytesseract>=0.3.13",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.17.0",
    "redis>=7.3.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "typer[standard]>=0.24.1",
    "uvicorn[standard]>=0.41.0",
]

[project.scripts]
boardgame = "boardgame_rules_backend.cli:app"
